Python/Ajedrez_IA/requirements.txt
Flask==3.0.0
python-chess==1.999